Added dpxStimRdkHuge, a derived class from dpxStimRdk that draws ovals instead of dots. This allows the dots to be used (the dot primitive has a maximum diameter in pixels that is determined by the hardware)
Class inheritance is awesome :). look into that dxpStimRdkHuge file, just 20 lines of simple code to solve this in a 100% backward compatible way!

checkaperture is now a proper member function of dpxStimRdk.m (needed to be inheritable)

Changed lkDpxTuningExp.m so that it uses the dpxStimHuge class when the speedDotDiam experiment is chosen. It could actually be used for all experiments, should be the same (except antialiasing, but we don't need that with the mice anyway).
Choose to use dxpStimRdkHuge only for this experiemnt only to not change the previous experiments at all.