MAJOR update to jdDpxExpHalfDomeRdkAnalysis/jdDpxExpHalfDomeRdkAnalysisSpeed.m
pooling of + and - speeds was incorrect. fixed now (briefly: first subtract bias (0-speed curve) and then flip the negative speed curves. then recalculate Mean trace and SEM etc

Added (ugly) information in the Figure about number of trials and mice that went into each curve. Not supposed to be printed in final figure of course. can be deleted and moved around using the white arrow tool in the matlab figure toolbar.

variable minRepsPerTracePerMouseToBeIncludedInMean on line 102 determines the minimum number of trials that a mouse must have performed on a particular condition for it to be included in the mean. Is now set to 30, works well.